+++ R/distr_beta_conc.R
@@ -1,11 +1,11 @@
 
-# BETA DISTRIBUTION / SHAPE PARAMETRIZATION
+# BETA DISTRIBUTION / CONCENTRATION PARAMETRIZATION
 
 
 # Parameters Function ----------------------------------------------------------
-distr_beta_shape_parameters <- function(n) {
-  group_of_par_names <- c("shape", "shape")
-  par_names <- c("shape", "shape")
+distr_beta_conc_parameters <- function(n) {
+  group_of_par_names <- c("conc", "conc")
+  par_names <- c("conc1", "conc2")
   par_support <- c("positive", "positive")
   res_parameters <- list(group_of_par_names = group_of_par_names, par_names = par_names, par_support = par_support)
   return(res_parameters)
@@ -14,7 +14,7 @@ distr_beta_shape_parameters <- function(n) {
 
 
 # Density Function -------------------------------------------------------------
-distr_beta_shape_density <- function(y, f) {
+distr_beta_conc_density <- function(y, f) {
   t <- nrow(f)
   a <- f[, 1, drop = FALSE]
   b <- f[, 2, drop = FALSE]
@@ -25,7 +25,7 @@ distr_beta_shape_density <- function(y, f) {
 
 
 # Log-Likelihood Function ------------------------------------------------------
-distr_beta_shape_loglik <- function(y, f) {
+distr_beta_conc_loglik <- function(y, f) {
   t <- nrow(f)
   a <- f[, 1, drop = FALSE]
   b <- f[, 2, drop = FALSE]
@@ -36,7 +36,7 @@ distr_beta_shape_loglik <- function(y, f) {
 
 
 # Mean Function ----------------------------------------------------------------
-distr_beta_shape_mean <- function(f) {
+distr_beta_conc_mean <- function(f) {
   t <- nrow(f)
   a <- f[, 1, drop = FALSE]
   b <- f[, 2, drop = FALSE]
@@ -47,7 +47,7 @@ distr_beta_shape_mean <- function(f) {
 
 
 # Variance Function ------------------------------------------------------------
-distr_beta_shape_var <- function(f) {
+distr_beta_conc_var <- function(f) {
   t <- nrow(f)
   a <- f[, 1, drop = FALSE]
   b <- f[, 2, drop = FALSE]
@@ -59,7 +59,7 @@ distr_beta_shape_var <- function(f) {
 
 
 # Score Function ---------------------------------------------------------------
-distr_beta_shape_score <- function(y, f) {
+distr_beta_conc_score <- function(y, f) {
   t <- nrow(f)
   a <- f[, 1, drop = FALSE]
   b <- f[, 2, drop = FALSE]
@@ -72,7 +72,7 @@ distr_beta_shape_score <- function(y, f) {
 
 
 # Fisher Information Function --------------------------------------------------
-distr_beta_shape_fisher <- function(f) {
+distr_beta_conc_fisher <- function(f) {
   t <- nrow(f)
   a <- f[, 1, drop = FALSE]
   b <- f[, 2, drop = FALSE]
@@ -87,7 +87,7 @@ distr_beta_shape_fisher <- function(f) {
 
 
 # Random Generation Function ---------------------------------------------------
-distr_beta_shape_random <- function(t, f) {
+distr_beta_conc_random <- function(t, f) {
   a <- f[1]
   b <- f[2]
   res_random <- be_silent(stats::dbeta(t, shape1 = a, shape2 = b))
@@ -98,7 +98,7 @@ distr_beta_shape_random <- function(t, f) {
 
 
 # Starting Estimates Function --------------------------------------------------
-distr_beta_shape_start <- function(y) {
+distr_beta_conc_start <- function(y) {
   y_mean <- mean(y, na.rm = TRUE)
   y_var <- stats::var(y, na.rm = TRUE)
   a <- max(y_mean * (y_mean * (1 - y_mean) / y_var - 1), 1e-6)