+++ data-raw/distr_table.R
@@ -4,7 +4,7 @@ library(tidyverse)
 library(readxl)
 
 distr_table <- read_excel("data-raw/distr_table.xlsx") %>%
-  mutate(type = factor(type, levels = c("binary", "categorical", "ranking", "count", "integer", "directional", "interval", "compositional", "duration", "real"))) %>%
+  mutate(type = factor(type, levels = c("binary", "cat", "ranking", "count", "integer", "circular", "interval", "comp", "duration", "real"))) %>%
   mutate(dim = factor(dim, levels = c("uni", "multi"))) %>%
   mutate(orthog = as.logical(orthog)) %>%
   mutate(default = as.logical(default)) %>%