+++ NEWS.md
@@ -1,7 +1,9 @@
 # gasmodel (development version)
 * Added diagonal and full Fisher information matrix scalings.
 * Added the Beta distribution.
+* Added the Dirichlet distribution.
 * Added the Laplace distribution.
+* Added the von Mises distribution.
 * Added german_car_market_cap dataset.
 * Corrected NA and Inf handling for the Plackett-Luce distribution.
 * Corrected the Fisher information for the generalized gamma distribution.